+++ tests/sentry/api/serializers/test_alert_rule_trigger_action.py
@@ -1,7 +1,7 @@
 from sentry.api.serializers import serialize
-from sentry.incidents.endpoints.serializers import action_target_type_to_string
 from sentry.incidents.logic import create_alert_rule_trigger, create_alert_rule_trigger_action
 from sentry.incidents.models import AlertRuleTriggerAction
+from sentry.incidents.serializers import ACTION_TARGET_TYPE_TO_STRING
 from sentry.testutils import TestCase
 
 
@@ -17,7 +17,7 @@ class AlertRuleTriggerActionSerializerTest(TestCase):
         )
         assert (
             result["targetType"]
-            == action_target_type_to_string[AlertRuleTriggerAction.TargetType(action.target_type)]
+            == ACTION_TARGET_TYPE_TO_STRING[AlertRuleTriggerAction.TargetType(action.target_type)]
         )
         assert result["targetIdentifier"] == action.target_identifier
         assert result["integrationId"] == action.integration_id
